\name{BostonHousingI}
\docType{data}
\alias{BostonHousingI}
\title{Boston Housing interaction data}
\description{
Housing data for 506 census tracts of Boston from the 1970 census.
This data set is modified to include all pairwise interactions of
main effects and to include B-spline basis functions of up to 6
degrees of freedom for all original predictors. The original data comprises 506 observations on 14 variables, \code{medv}
being the outcome variable. Other variables represents original
variables and their interaction variables which includes altogether
148 covariates.

}

\examples{data(BostonHousingI, package = "modelSampler")}
\keyword{datasets}
