\name{OzoneI}
\docType{data}
\alias{OzoneI}
\title{Ozone interaction data}
\description{
Original Ozone data set is modified to include
all pairwise interactions of main effects and to include B-spline
basis functions of up to 6 degrees of freedom for all original
predictors. The original data comprises 203 observations on 14 variables, \dQuote{ozone}
being the outcome variable. Other variables represents original
variables and their interaction variables which includes altogether
89 covariates.

}



\examples{data(OzoneI, package = "modelSampler")}
\keyword{datasets}

